Fixed the dependency issues and created the entity relations

+++ backend/src/main/java/com/library/repository/LoanRepository.java
@@ -4,6 +4,8 @@ import com.library.model.Loan;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface LoanRepository extends CrudRepository<Loan, Long> {
+public interface LoanRepository extends CrudRepository<Loan, BigInteger> {
 }
