+++ backend/src/main/java/com/library/repository/UserRepository.java
@@ -4,6 +4,8 @@ import com.library.model.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, BigInteger> {
 }
